requirements.txt
flask
openai
pillow
SpeechRecognition
pyaudio